YT-3419: Change TMessageId from int32 to int64

+++ yt/ytlib/hive/public.h
@@ -21,7 +21,7 @@ class TEncapsulatedMessage;
 
 ////////////////////////////////////////////////////////////////////////////////
 
-using TMessageId = int;
+using TMessageId = i64;
 
 struct TCellDescriptor;
 struct TCellInfo;
